Added file serialization, users, and testing
Major:
Added the ability to save and load media and media sets, so that favorite libraries can be saved and loaded in the future.

Added the user class together with simple XOR-encryption for passwords. (Because why not)

Added bunch of tests for file parsing and such.

Minor:
Separated MediaSet from MediaLibrary and put it in the Data package.

Separated Movie and Serie into their own files.